pyproject.toml
[tool.poetry]
name = "tavern-flask"
version = "0.1.0"
description = "Flask plugin for tavern"
authors = ["Jérémy Tellaa <jeremy.tellaa@gmail.com>"]
readme = "README.md"
packages = [{include = "tavern_flask"}]

[tool.poetry.dependencies]
python = "^3.11"
tavern = "^2.3"
PyYAML = "^6.0"
Flask = "^3.0"
python-box = "^6.1"
requests = "^2.31"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1"
isort = "^5.12"
black = "^23.9"
pytest = "^7.2"
pytest-cov = "^4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."tavern_http"]
"flask" = "tavern_flask.tavernhook"
